fix: use px for breakpoints
Fixes a bug where responsive type and REM breakpoints was leading to
multiple breakpoints being active at once.

Also fixed issue on medium breakpoint where the top of the range was the
same as the bottom of the large range. Now it subtracts 1 pixel from
this value to ensure no overlap.

The px values were chosen based on where our current values are, but
these may not be the most common values so we can adjust. I wanted to
have the least impact.